requirements.txt
Django==4.1.4
django-environ==0.9.0
psycopg2==2.9.2
Pillow==9.0.1
django-phonenumber-field==1.3.0
pip==22.3.1
requests==2.8.1
simplejson==3.18.0
django-crum==0.7.9
django-crispy-forms==2.0
crispy-bootstrap4
matplotlib==3.4.3
reportlab==3.6.1
